Maybe the last update
Maybe the last update

+++ renren-security/renren-admin/src/main/java/io/renren/modules/sys/entity/LzConsumptionEntity.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -31,7 +32,7 @@ public class LzConsumptionEntity implements Serializable {
 	/**
 	 * 消费金额
 	 */
-	private Float money;
+	private BigDecimal money;
 	/**
 	 * 消费时间
 	 */
